Provide reusable helper classes for currently intended functionality
